[ALPS05182223] lcm driver:k85 coverity issue fix
[Detail] ALPS05182223 CERT-C Characters and Strings:10615315

[Solution] fix it

MTK-Commit-Id: b194518f3e28a24b8e42efd83bd9492eeb248df7

CR-Id: ALPS05182223
Feature: [Module]Display Driver

+++ drivers/misc/mediatek/video/mt6785/dispsys/display_recorder.c
@@ -1358,7 +1358,7 @@ char *get_dprec_status_ptr(int buffer_idx)
 
 static char *_logger_pr_type_spy(enum DPREC_LOGGER_PR_TYPE type)
 {
-	if (type >= DPREC_LOGGER_PR_NUM)
+	if ((unsigned int)type >= DPREC_LOGGER_PR_NUM)
 		return "unknown";
 	else
 		return dprec_logger_buffer[type].name;
